+++ packages/studio-explore/src/services/cypher/queryStatistics.tsx
@@ -14,7 +14,6 @@ export const queryStatistics = async () => {
     if (engine === 'graphscope') {
       // const script = `call gs.procedure.meta.statistics()`;
       // const result = await queryStatement(script);
-      // debugger;
       // console.log('result', result, result.raw.records[0]._fields[0]);
       // return JSON.parse(result.raw.records[0]._fields[0]);
     }