+++ packages/studio-driver/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphscope/studio-driver
 
+## 0.1.12
+
+### Patch Changes
+
+- update studio graph
+
 ## 0.1.11
 
 ### Patch Changes