+++ packages/studio-graph/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @graphscope/studio-graph
 
+## 0.1.14
+
+### Patch Changes
+
+- update studio graph
+- Updated dependencies
+  - @graphscope/studio-components@0.1.14
+  - @graphscope/use-zustand@0.1.6
+
 ## 0.1.13
 
 ### Patch Changes